multiple: refactor prepareWorkingImage and common command options
